+++ projects/angular2-smart-table/src/lib/components/thead/cells/add-button.component.ts
@@ -62,7 +62,7 @@ export class AddButtonComponent implements AfterViewInit, OnChanges {
         source: this.source,
       });
     } else {
-      this.grid.createFormShown = true;
+      this.grid.showCreateForm();
     }
   }
 }